requirements.txt
pandas
numpy
requests
pyodbc
python-dotenv
mysql-connector-python
seaborn==0.12.2
matplotlib==3.7.1
SQLAlchemy==2.0.19
PyMySQL==1.0.3

# MySQL dependencies
mysql-connector-python==8.0.33

# Logging
loguru==0.6.0       # Enhanced logging

# Testing
pytest==7.4.0       # Testing framework

# Additional dependencies (if needed for ETL tasks)
python-dotenv==1.0.0  # Load environment variables
